Relocate AllowancesTable and fix styles
