+++ src/components/Office/MoveDetailTable.module.scss
@@ -3,7 +3,7 @@
 @import 'shared/styles/colors';
 @import 'shared/styles/_custom';
 
-.OrdersTable {
+.MoveDetailTable {
   @include u-padding-y(2);
   :global .stackedtable-header {
     max-width: unset;
